change routes and header
