+++ scss/style.scss
@@ -1,7 +1,7 @@
 @import "../node_modules/bootstrap/scss/bootstrap";
 
 body {
-  padding-top: 80px;
+  padding-top:200px;
   min-height: 900px;
   background-image: linear-gradient(290deg, #fd8003,#504aa3, #8a1253);
